Feature: added add activity to profile
