+++ src/Components/Activity/Individual/SingleAct.jsx
@@ -1,0 +1,9 @@
+import RequestCard from "./RequestCard";
+
+export default function SingleAct() {
+  return (
+    <div>
+      <RequestCard />
+    </div>
+  );
+}
